requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
elevenlabs>=1.0.0
pygame>=2.6.0
